# BackEnd/Swagger.json
{"openapi":"3.1.0","info":{"title":"LicitSystem API","description":"API para gerenciamento de licitações e anexos","version":"1.0.0"},"paths":{"/departments/":{"get":{"tags":["Departments"],"summary":"Get All Departments","operationId":"get_all_departments_departments__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Departments"],"summary":"Create Department","operationId":"create_department_departments__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/departments/{department_id}":{"get":{"tags":["Departments"],"summary":"Get Department By Id","operationId":"get_department_by_id_departments__department_id__get","parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"integer","title":"Department Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Departments"],"summary":"Update Department","operationId":"update_department_departments__department_id__patch","parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"integer","title":"Department Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Departments"],"summary":"Delete Department","operationId":"delete_department_departments__department_id__delete","parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"integer","title":"Department Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/categories/":{"get":{"tags":["Categories"],"summary":"Get All Categories","operationId":"get_all_categories_categories__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Categories"],"summary":"Create Category","operationId":"create_category_categories__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/categories/{category_id}":{"get":{"tags":["Categories"],"summary":"Get Category By Id","operationId":"get_category_by_id_categories__category_id__get","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","title":"Category Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Categories"],"summary":"Update Category","operationId":"update_category_categories__category_id__patch","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","title":"Category Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Categories"],"summary":"Delete Category","operationId":"delete_category_categories__category_id__delete","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","title":"Category Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/biddings/":{"get":{"tags":["Biddings"],"summary":"Get All Biddings","operationId":"get_all_biddings_biddings__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}},{"name":"number","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Number"}},{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"}},{"name":"department_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Department Id"}},{"name":"category_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Category Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Biddings"],"summary":"Create Bidding","operationId":"create_bidding_biddings__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BiddingCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/biddings/{bidding_id}":{"get":{"tags":["Biddings"],"summary":"Get Bidding By Id","operationId":"get_bidding_by_id_biddings__bidding_id__get","parameters":[{"name":"bidding_id","in":"path","required":true,"schema":{"type":"integer","title":"Bidding Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Biddings"],"summary":"Update Bidding","operationId":"update_bidding_biddings__bidding_id__patch","parameters":[{"name":"bidding_id","in":"path","required":true,"schema":{"type":"integer","title":"Bidding Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BiddingUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Biddings"],"summary":"Delete Bidding","operationId":"delete_bidding_biddings__bidding_id__delete","parameters":[{"name":"bidding_id","in":"path","required":true,"schema":{"type":"integer","title":"Bidding Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licitacoes/{licitacao_id}/anexos":{"post":{"tags":["Anexos"],"summary":"Upload Attachment","operationId":"upload_attachment_licitacoes__licitacao_id__anexos_post","parameters":[{"name":"licitacao_id","in":"path","required":true,"schema":{"type":"integer","title":"Licitacao Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_attachment_licitacoes__licitacao_id__anexos_post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Anexos"],"summary":"List Attachments","operationId":"list_attachments_licitacoes__licitacao_id__anexos_get","parameters":[{"name":"licitacao_id","in":"path","required":true,"schema":{"type":"integer","title":"Licitacao Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttachmentResponse"},"title":"Response List Attachments Licitacoes  Licitacao Id  Anexos Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/anexos/{attachment_id}/download":{"get":{"tags":["Anexos"],"summary":"Download Attachment","operationId":"download_attachment_anexos__attachment_id__download_get","parameters":[{"name":"attachment_id","in":"path","required":true,"schema":{"type":"integer","title":"Attachment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/anexos/{attachment_id}":{"delete":{"tags":["Anexos"],"summary":"Delete Attachment","operationId":"delete_attachment_anexos__attachment_id__delete","parameters":[{"name":"attachment_id","in":"path","required":true,"schema":{"type":"integer","title":"Attachment Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard/":{"get":{"tags":["Dashboard"],"summary":"Get Summary","operationId":"get_summary_dashboard__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard/biddings":{"get":{"tags":["Dashboard"],"summary":"Get Latest Biddings","operationId":"get_latest_biddings_dashboard_biddings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AttachmentResponse":{"properties":{"id":{"type":"integer","title":"Id"},"licitacao_id":{"type":"integer","title":"Licitacao Id"},"nome":{"type":"string","title":"Nome"},"caminho":{"type":"string","title":"Caminho"},"tipo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tipo"},"categoria":{"type":"string","title":"Categoria"},"tamanho_kb":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tamanho Kb"},"criado_em":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Criado Em"}},"type":"object","required":["id","licitacao_id","nome","caminho","categoria"],"title":"AttachmentResponse"},"BiddingCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"department_id":{"type":"integer","title":"Department Id"},"category_id":{"type":"integer","title":"Category Id"},"number":{"type":"integer","title":"Number"},"year":{"type":"integer","title":"Year"},"bidding_type":{"type":"string","title":"Bidding Type"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","default":"Aberto"},"classification":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Classification","default":"Global"},"object_name":{"type":"string","title":"Object Name"},"object_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Object Description"},"estimated_value":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Estimated Value"},"publication_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Publication Date"},"opening_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Opening Date"}},"type":"object","required":["user_id","department_id","category_id","number","year","bidding_type","object_name"],"title":"BiddingCreate"},"BiddingUpdate":{"properties":{"department_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Department Id"},"category_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Category Id"},"number":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Number"},"year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"},"bidding_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bidding Type"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"classification":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Classification"},"object_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Object Name"},"object_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Object Description"},"estimated_value":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Estimated Value"},"publication_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Publication Date"},"opening_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Opening Date"}},"type":"object","title":"BiddingUpdate"},"Body_upload_attachment_licitacoes__licitacao_id__anexos_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_attachment_licitacoes__licitacao_id__anexos_post"},"CategoryCreate":{"properties":{"nome":{"type":"string","title":"Nome"},"tipo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tipo","default":"Comum"}},"type":"object","required":["nome"],"title":"CategoryCreate"},"CategoryUpdate":{"properties":{"nome":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nome"},"tipo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tipo"}},"type":"object","title":"CategoryUpdate"},"DepartmentCreate":{"properties":{"sigla":{"type":"string","title":"Sigla"},"nome":{"type":"string","title":"Nome"}},"type":"object","required":["sigla","nome"],"title":"DepartmentCreate"},"DepartmentUpdate":{"properties":{"sigla":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sigla"},"nome":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nome"}},"type":"object","title":"DepartmentUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}